Move constraints-related bindable test cases to `ConstrainedBindableTest`
